Moved ResolvedIdentifierResultHolder to it's own file.
